fix Fluent Macros
https://wiki.ralfbarkow.ch/view/fluent-macros

We've created javascript classes that wrap the nodes and rels we use to represent a graph. This is further wrapped with Nodes and Rels classes that represent subsets of the graph and offer fluent query functions to traverse the graph based on types, properties or just the availability of relations.