Enrich Any SVG
We look for all SVG files in lineup assets including the copy at the top of this page. Click on the file SVG file name below and a FedWiki page will be created with a clickable SVG in it.

http://marc.tries.fed.wiki/view/enrich-any-svg

I see a problem in the convert-any-svg script.

Added console.info("convert-any-svg.html: url: ", url) to see which url is used.