Enrich Any SVG ⇒ Converted SVG
Revert console.info change.

The issue with the convert-any-svg script probably had to do with a misunderstanding on my part. As soon as I uploaded one of Marc's svg files to wiki.ralfbarkow.ch, it worked. Since the script also finds files that are only present on the marc.tries. fed.wiki server, I thought the script had magic powers.

https://wiki.ralfbarkow.ch/view/enrich-any-svg/view/converted-svg